+++ resources/views/projeto/orcamento/index.blade.php
@@ -22,7 +22,9 @@
                 <div class="card-body">
                     <div class="table-header">
                         <h4 class="header-title">Lista de Orçamentos do Projeto: <b>{{ $projeto->titulo }}</b></h4>
-                        <a href="{{url()->current()}}/create" class="btn btn-success btn-create">Cadastrar</a>
+                        @can('projeto.orcamento.create')
+                            <a href="{{url()->current()}}/create" class="btn btn-success btn-create">Cadastrar</a>
+                        @endcan
                     </div>
                     <p class="text-muted fs-14">
                         Abaixo a lista com todos os orçamentos.
@@ -49,14 +51,18 @@
                                             @endif
                                         @endforeach
                                         <td>
-                                            <div style="display: inline-block" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="primary-tooltip" data-bs-title="Editar registro">
-                                                <a href="{{url()->current()}}/edit/{{$item->id}}" class="text-reset px-1" style="font-size: 20px"> <i class="ri-pencil-line"></i></a>
-                                            </div>                                      
-                                            <div style="display: inline-block" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="danger-tooltip" data-bs-title="Excluir registro">
-                                                <a href="javascript: deleteModal({{ $item->id }});" class="text-reset px-1" style="font-size: 20px"> 
-                                                    <i class="ri-delete-bin-2-line"></i>
-                                                </a>
-                                            </div>                                            
+                                            @can('projeto.orcamento.update')
+                                                <div style="display: inline-block" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="primary-tooltip" data-bs-title="Editar registro">
+                                                    <a href="{{url()->current()}}/edit/{{$item->id}}" class="text-reset px-1" style="font-size: 20px"> <i class="ri-pencil-line"></i></a>
+                                                </div>
+                                            @endcan
+                                            @can('projeto.orcamento.delete')                                    
+                                                <div style="display: inline-block" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-custom-class="danger-tooltip" data-bs-title="Excluir registro">
+                                                    <a href="javascript: deleteModal({{ $item->id }});" class="text-reset px-1" style="font-size: 20px"> 
+                                                        <i class="ri-delete-bin-2-line"></i>
+                                                    </a>
+                                                </div>
+                                            @endcan                                            
                                         </td>
                                     </tr>
                                 @endforeach
