+++ routes/web.php
@@ -26,32 +26,31 @@ Route::middleware('auth')->group(function() {
 
     // Projeto
     Route::prefix('projetos')->group(function (){
-        Route::post('/', [ProjetoController::class, 'store'])->name('projeto.store');
-        Route::get('/', [ProjetoController::class, 'index'])->name('projeto.index');
-        Route::get('/create', [ProjetoController::class, 'create'])->name('projeto.create');
-        Route::get('/edit/{id}', [ProjetoController::class, 'edit'])->name('projeto.edit');
-        Route::patch('/edit/{id}', [ProjetoController::class, 'update'])->name('projeto.update');
-        Route::delete('/{id}', [ProjetoController::class, 'destroy'])->name('projeto.destroy');
+        Route::get('/', [ProjetoController::class, 'index'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.read')])->name('projeto.index');
+        Route::post('/', [ProjetoController::class, 'store'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.create')])->name('projeto.store');
+        Route::get('/create', [ProjetoController::class, 'create'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.create')])->name('projeto.create');
+        Route::get('/edit/{id}', [ProjetoController::class, 'edit'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.update')])->name('projeto.edit');
+        Route::patch('/edit/{id}', [ProjetoController::class, 'update'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.update')])->name('projeto.update');
+        Route::delete('/{id}', [ProjetoController::class, 'destroy'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.delete')])->name('projeto.destroy');
 
         Route::prefix('/{id_projeto}/orcamentos')->group(function (){
-            Route::post('/', [ProjetoOrcamentoController::class, 'store'])->name('projeto.orcamento.store');
-            Route::get('/', [ProjetoOrcamentoController::class, 'index'])->name('projeto.orcamento.index');
-            Route::get('/create', [ProjetoOrcamentoController::class, 'create'])->name('projeto.orcamento.create');
-            Route::get('/edit/{id}', [ProjetoOrcamentoController::class, 'edit'])->name('projeto.orcamento.edit');
-            Route::patch('/edit/{id}', [ProjetoOrcamentoController::class, 'update'])->name('projeto.orcamento.update');
-            Route::delete('/{id}', [ProjetoOrcamentoController::class, 'destroy'])->name('projeto.orcamento.destroy');
+            Route::get('/', [ProjetoOrcamentoController::class, 'index'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.orcamento.read')])->name('projeto.orcamento.index');
+            Route::post('/', [ProjetoOrcamentoController::class, 'store'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.orcamento.create')])->name('projeto.orcamento.store');
+            Route::get('/create', [ProjetoOrcamentoController::class, 'create'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.orcamento.create')])->name('projeto.orcamento.create');
+            Route::get('/edit/{id}', [ProjetoOrcamentoController::class, 'edit'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.orcamento.update')])->name('projeto.orcamento.edit');
+            Route::patch('/edit/{id}', [ProjetoOrcamentoController::class, 'update'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.orcamento.update')])->name('projeto.orcamento.update');
+            Route::delete('/{id}', [ProjetoOrcamentoController::class, 'destroy'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('projeto.orcamento.delete')])->name('projeto.orcamento.destroy');
         });
-
     });
 
     // Users routes
     Route::prefix('users')->group(function () {
-        Route::post('/', [UserController::class, 'store'])->name('user.store');
-        Route::get('/', [UserController::class, 'index'])->name('user.index');
-        Route::get('/create', [UserController::class, 'create'])->name('user.create');
-        Route::get('/edit/{id}', [UserController::class, 'edit'])->name('user.edit');
-        Route::patch('/edit/{id}', [UserController::class, 'update'])->name('user.update');
-        Route::delete('/{id}', [UserController::class, 'destroy'])->name('user.destroy');
+        Route::get('/', [UserController::class, 'index'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('user.read')])->name('user.index');
+        Route::post('/', [UserController::class, 'store'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('user.create')])->name('user.store');
+        Route::get('/create', [UserController::class, 'create'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('user.create')])->name('user.create');
+        Route::get('/edit/{id}', [UserController::class, 'edit'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('user.update')])->name('user.edit');
+        Route::patch('/edit/{id}', [UserController::class, 'update'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('user.update')])->name('user.update');
+        Route::delete('/{id}', [UserController::class, 'destroy'])->middleware([\Illuminate\Auth\Middleware\Authorize::using('user.delete')])->name('user.destroy');
     });
 });
